feat: Return metadata response even if getting Trivy version command fails
It might happen that Trivy DB is corrupted and trivy -version fails.
Then the /api/v1/matadata response returns 500 status code.
However the /api/v1/metadata endpoint is used by Harbor as a health check.
The only way to fix the Trivy DB is to rescan an image, but with
a fialing health check it's not possible.

This commit returns /api/v1/metadata with 200 status code even if
getting trivy -version failes.

Signed-off-by: Daniel Pacak <pacak.daniel@gmail.com>